requirements.txt
cffi==2.0.0
numpy==2.4.4
pycparser==3.0
pyserial==3.5
sounddevice==0.5.5
